+++ src/main/java/ru/stegnin/virtualbox/api/controller/UserController.java
@@ -1,4 +1,4 @@
-package ru.stegnin.virtualbox.controller;
+package ru.stegnin.virtualbox.api.controller;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -6,11 +6,11 @@ import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
-import ru.stegnin.virtualbox.model.User;
-import ru.stegnin.virtualbox.repository.AuthRepository;
-import ru.stegnin.virtualbox.repository.UserRepository;
-import ru.stegnin.virtualbox.support.Constants;
-import ru.stegnin.virtualbox.support.GenericResponse;
+import ru.stegnin.virtualbox.api.model.User;
+import ru.stegnin.virtualbox.api.repository.AuthRepository;
+import ru.stegnin.virtualbox.api.repository.UserRepository;
+import ru.stegnin.virtualbox.settings.support.Constants;
+import ru.stegnin.virtualbox.settings.support.GenericResponse;
 
 import java.util.ArrayList;
 import java.util.List;
