+++ src/main/java/ru/stegnin/virtualbox/api/repository/AuthRepository.java
@@ -1,0 +1,7 @@
+package ru.stegnin.virtualbox.api.repository;
+
+import java.net.URI;
+
+public interface AuthRepository {
+    URI getUri(Class cl, String methodName, Object... args);
+}
