+++ src/main/java/ru/stegnin/virtualbox/settings/security/KeyGenerator.java
@@ -1,4 +1,4 @@
-package ru.stegnin.virtualbox.security;
+package ru.stegnin.virtualbox.settings.security;
 
 import java.security.Key;
